requirements.txt
aiogram==2.25
psycopg2-binary==2.9.5
aiohttp==3.8.4
apscheduler==3.10.1
